Improve pushpkg and update_repo
Add delete capability to pushpkg
Add c-pod base specifier for non-default user

+++ bin/rx_pkg.cgi
@@ -11,12 +11,18 @@ File.umask(0002)
 $repo = File.absolute_path('..', File.dirname(__FILE__))
 $base = File.absolute_path('../..', File.dirname(__FILE__))
 
-def repopath pkgname, type
+def warn msg
+    "\033[31m#{msg}\033[0m"
+end
+
+# Return the path on disk for this package and the base name
+#
+def repopath package, type
     type = 'unstable' if type == ''
-    raise "#{pkgname} isn't a recognizable package" unless /(?<package>^.*)-(?<pkginfo>.*)$/ =~ pkgname
+    raise "Can't parse #{package}" unless /(?<pkgname>^.*)-(?<ver>\d[0-9\.]+\d)(?<pkginfo>.*)$/ =~ package
     pkgparts = pkginfo.split '.'
     suffix = pkgparts.pop
-    case suffix
+    path = case suffix
     when 'rpm'
 	arch = pkgparts.pop
 	rel = if pkgparts.select { |c| c =~ /^(rh)?el5(_.*)?$/ }.size > 0
@@ -28,14 +34,27 @@ def repopath pkgname, type
 	       end
 	yum_repos = File.absolute_path('../../yum_repos', File.dirname(__FILE__))
 	if ['stable','unstable'].include? type
-	    "#{$base}/yum_repos/custom/#{rel}/#{type}/#{arch}/#{pkgname}"
+	    "#{$base}/yum_repos/custom/#{rel}/#{type}/#{arch}"
 	else
-	    "#{$base}/yum_repos/#{type}/#{rel}/#{arch}/#{pkgname}"
+	    "#{$base}/yum_repos/#{type}/#{rel}/#{arch}"
 	end
     when 'gem'
-	"#{$base}/gem_repo/gems/#{pkgname}"
+	"#{$base}/gem_repo/gems"
     else raise "Suffix #{suffix} is not a recognized package type"
     end
+    return path, pkgname
+end
+
+# Retain only the latest copies of the packages
+#
+def cleanup path, name, copies
+    return "" if copies == -1
+    files = Dir.glob("#{path}/#{name}*")
+    files.select!{|f| /(?<pkgname>^.*)-(?<ver>\d[0-9\.]+\d)(?<pkginfo>.*)$/ =~ File.basename(f) and pkgname == name }
+    files.sort!{|a,b| File.mtime(b) <=> File.mtime(a)}
+    files.shift(copies - 1) if copies > 1
+    files.each{|f| File.unlink f} 
+    return warn "Deleted #{files.size} older versions of #{name} in path #{path}\n"
 end
 
 q = CGI.new
@@ -45,18 +64,25 @@ begin
     if q.include?('pkgfile')
 	t = q['pkgfile']
 	filename = t.original_filename
-	ofile = repopath(filename, q['type'])
-	case t
-	when StringIO
-	    IO.copy_stream t, ofile
-	when Tempfile
-	    FileUtils.cp t.local_path, ofile
-	    t.unlink
-	else
-	    raise "I unexpectedly got #{t.inspect}"
-	end
-	FileUtils.chmod 0660, ofile
-	q.print "File uploaded to #{ofile}\n"
+	opath, pkgname = repopath(filename, q['type'])
+        copies = q.include?('keep') ? q['keep'].to_i : -1
+        delmsg = cleanup opath, pkgname, copies
+        if copies == 0
+            q.print warn "Deleted all versions of #{pkgname} in path #{opath}\n"
+        else
+            ofile = "#{opath}/#{filename}"
+            case t
+            when StringIO
+                IO.copy_stream t, ofile
+            when Tempfile
+                FileUtils.cp t.local_path, ofile
+                t.unlink
+            else
+                raise "I unexpectedly got #{t.inspect}"
+            end
+            FileUtils.chmod 0660, ofile
+            q.print "Uploaded to #{ofile}\n#{delmsg}"
+        end
     end
     if q['rebuild'] == 'yes'
 	rebuild_opts = q['force'] == "yes" ? "-f": ""
